deduct ap on movement
