+++ module/falloutzero.mjs
@@ -3,13 +3,11 @@ import * as models from './data/_module.mjs'
 import * as documents from './documents/_module.mjs'
 import * as sheets from './sheets/_module.mjs'
 import { preloadHandlebarsTemplates } from './helpers/templates.mjs'
-import SkillRoll from './dice/skill-roll.mjs'
-import { allowMovement } from './helpers/movement.mjs'
-import FalloutZeroArmor from './data/armor.mjs'
-import FalloutZeroItem from './documents/item.mjs'
+import { registerSystemSettings, registerHbsHelpers } from './registerSettings.mjs'
 
 // Import Submodules
 import * as applications from '../module/applications/_module.mjs'
+import { registerHooks } from './registerHooks.mjs'
 
 /* -------------------------------------------- */
 /*  Init Hook                                   */
@@ -22,78 +20,9 @@ Hooks.once('init', function () {
     applications,
     rollItemMacro,
   }
-  game.settings.register('core', 'CarryLoad', {
-    name: 'Exact Carry Load Calculator',
-    hint: 'Checked: 23 x 10mm ammo = 2.3 load | Unchecked: 23 x 10mm ammo = 2 load',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: false,
-    requiresReload: true,
-  })
-  game.settings.register('core', 'CapsLoad', {
-    name: 'Do Caps Have Load?',
-    hint: 'Checked: 50 caps is 1 Load',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: true,
-    requiresReload: true,
-  })
-  game.settings.register('core', 'AmmoLoad', {
-    name: 'Does Ammo Have Load?',
-    hint: 'Checked: 10 Ammo is 1 load',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: true,
-    requiresReload: true,
-  })
-  game.settings.register('core', 'JunkLoad', {
-    name: 'Does Junk Have Load?',
-    hint: 'If Checked: 5 Junk is 1 Load, 10 Material is 1 Load',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: true,
-    requiresReload: true,
-  })
-  game.settings.register('core', 'Sheet-Color', {
-    name: 'Your Sheet Theme Color',
-    hint: 'An override for sheet color. Green = #1bff80 / Amber = #ffb641 or you can use custom hex or just say "red" or "green" Blank puts it back to User Color being used.',
-    scope: 'client',
-    config: true,
-    type: String,
-    default: '',
-    requiresReload: true,
-  })
-  game.settings.register('core', 'VaultTec', {
-    name: 'Use Vault-Tec Colors',
-    hint: 'An Alternate Color Scheme Sponsored By Vault Tec!',
-    scope: 'client',
-    config: true,
-    type: Boolean,
-    default: false,
-    requiresReload: true,
-  })
-  game.settings.register('core', 'DeductMovementAPInCombat', {
-    name: 'Auto deduct movement AP',
-    hint: 'Automatically reduces AP based on character movement when in combat. [Requires Elevation Ruler]',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: true,
-    requiresReload: true,
-  })
-  game.settings.register('core', 'AutoRecycleAP', {
-    name: 'Auto recycle AP',
-    hint: 'Automatically recycles AP when combat round advances',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: true,
-    requiresReload: true,
-  })
+
+  // Registers system settings
+  registerSystemSettings()
 
   const sheetColor = game.settings.get('core', 'Sheet-Color');
   const r = document.querySelector(':root');
@@ -186,282 +115,12 @@ Hooks.once('init', function () {
 /* -------------------------------------------- */
 /*  Handlebars Helpers                          */
 /* -------------------------------------------- */
-
-// If you need to add Handlebars helpers, here is a useful example:
-Handlebars.registerHelper('toLowerCase', function (str) {
-  return str.toLowerCase()
-})
-
-Handlebars.registerHelper('Reload', function (v1) {
-  if (v1.includes("Manual Reload")) {
-    return 1
-  }
-  if (v1.includes("Quick Reload")) {
-    return 4
-  }
-  return 6
-})
-
-Handlebars.registerHelper('log', function (v1) {
-  console.log(v1)
-})
-
-Handlebars.registerHelper('setChecked', function (value, test) {
-  if (value == undefined) return ''
-  return value == test ? 'checked' : ''
-})
-
-// If Player is a GM
-Handlebars.registerHelper('GM', function (options) {
-  if (game.user.role === 4) {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-
-// If Character is a NPC
-Handlebars.registerHelper('NPC', function (actorType, options) {
-  if (actorType == 'npc') {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-// If Vault Tec Sheets
-Handlebars.registerHelper('VaultTec', function (actorType, options) {
-  if (game.settings.get('core', 'VaultTec')) {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-// If Value equals something
-Handlebars.registerHelper('Check', function (v1, v2, options) {
-  if (v1 == v2) {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-
-Handlebars.registerHelper('IsIn', function (item, list) {
-  if (list.includes(item)) {
-    return true
-  }
-  return false
-})
-
-// Greater Than or Equal
-Handlebars.registerHelper('GreaterThan', function (v1, v2, options) {
-  if (v1 >= v2) {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-// Less Than
-Handlebars.registerHelper('LesserThan', function (v1, v2, options) {
-  if (v1 < v2) {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-
-// Less Than
-Handlebars.registerHelper('DifferentFrom', function (v1, v2, options) {
-  if (v1 != v2) {
-    return options.fn(this)
-  }
-  return options.inverse(this)
-})
-
-// Sum Of
-Handlebars.registerHelper('Sum', function (v1, v2) {
-  let sum = Number(v1) + Number(v2)
-  return sum
-})
-Handlebars.registerHelper('Sum3', function (v1, v2, v3) {
-  let sum = Number(v1) + Number(v2) + Number(v3)
-  return sum
-})
-Handlebars.registerHelper('Subtract', function (v1, v2) {
-  let subtract = Number(v1) - Number(v2)
-  return subtract
-})
-Handlebars.registerHelper('Skills', function (v1, v2, v3, v4) {
-  let sum = Number(v1) + Number(v2) + Number(v3) - Number(v4)
-  return sum
-})
-//division
-Handlebars.registerHelper('LckMod', function (v1, v2) {
-  let div = Math.floor(Number(v1) / Number(v2))
-  if (div < -1) {
-    div = -1
-  }
-  return div
-})
-//multiplication
-Handlebars.registerHelper('Multiply', function (v1, v2) {
-    let mathResult = Math.floor(Number(v1) * Number(v2))
-    return mathResult
-})
-
-Handlebars.registerHelper('NotEqual', function (v1, v2) {
-  return v1 !== v2
-})
-
-Handlebars.registerHelper('Equals', function (v1, v2) {
-  return v1 === v2
-})
-
-//Format a Compendium Link for a given title
-Handlebars.registerHelper('FormatCompendium', function (itemName, compendium) {
-  let compendiumObject, myItem
-  try {
-    compendiumObject = game.packs.find((u) => u.metadata.name == compendium)
-    myItem = compendiumObject.tree.entries.find(
-      (u) => u.name.toLowerCase() == itemName.toLowerCase(),
-    )
-    if (myItem) {
-      return `<a class="content-link"  draggable="true" data-link data-uuid="Compendium.arcane-arcade-fallout.${compendium}.Item.${myItem._id}" 
-        data-id="${myItem._id}" data-type="Item" data-pack="arcane-arcade-fallout.${compendium}">
-        ${itemName}</a>`
-    } else {
-      return `${itemName}`
-    }
-  } catch {
-    return `${itemName}`
-  }
-})
-
-Handlebars.registerHelper('SpecialKeyToLabel', function (key) {
-  return FALLOUTZERO.abilities[key].label || ''
-})
+registerHbsHelpers()
 
 /* -------------------------------------------- */
-/*  Ready Hook                                  */
+/*  Hooks                                  */
 /* -------------------------------------------- */
-
-Hooks.once('ready', function () {
-  // Wait to register hotbar drop hook on ready so that modules could register earlier if they want to
-  Hooks.on('hotbarDrop', (bar, data, slot) => createItemMacro(data, slot))
-
-  // Auto recycle AP on turn end
-  if (game.user.isGM && game.settings.get('core', 'AutoRecycleAP')) {
-    Hooks.on("updateCombat", async (combat, updates, update) => {
-      // if round did not change or direction is backwards, return
-      if (!updates.round || update.direction !== 1) return
-
-      // else recycle ap for all combatants
-      game.combat.combatants.forEach((combatant) => {
-        combatant.actor.recycleAp()
-      })
-    });
-  }
-})
-
-/* --------------------------------------------  */
-/*  Token movement                                 */
-/* --------------------------------------------  */
-// block movement if not turn
-Hooks.on('preUpdateToken', (document, update) => {
-  if (!game.settings.get('core', 'DeductMovementAPInCombat')) return
-  if ((update.x != undefined || update.y != undefined) && game.combats.active) {
-    let allow = allowMovement(document);
-
-    const ruler = canvas.controls.ruler;
-    let lastMoveDistance = 0;
-    if ( ruler.active && ruler.token === _token ) {
-      // Ruler move
-      lastMoveDistance = ruler.totalCost - ruler.history.reduce((acc, curr) => acc + curr.cost, 0);
-      // numDiagonal = ruler.totalDiagonals;
-    }
-    // else {
-    //   // Some other move; likely arrow keys.
-    //   const numPrevDiagonal = game.combat?.started ? (token._combatMoveData?.numDiagonal ?? 0) : 0;
-    //   const mp = new MovePenalty(token);
-    //   const res = mp.measureSegment(token.position, token.document._source, { numPrevDiagonal });
-    //   lastMoveDistance = res.cost;
-    //   numDiagonal = res.numDiagonal;
-    // }
-
-    const distanceCost = lastMoveDistance / 5
-    allow = document.actor.applyApCost(distanceCost)
-
-    if (!allow) {
-        delete update.x;
-        delete update.y;
-    }
-  }
-});
-
-// deduct AP on move in combat
-// Hooks.on('updateToken', (document, update) => {
-//   if (!game.combats.active) return
-
-//   if ((update.x != undefined || update.y != undefined)) {
-//     if (_token?.lastMoveDistance !== undefined) {
-//       console.log('distance', _token.lastMoveDistance)
-//       console.log('cost', _token.lastMoveDistance / 5)
-//       document.actor.applyApCost(_token.lastMoveDistance / 5)
-//     }
-//   }
-// });
-
-/* --------------------------------------------  */
-/*  Other Hooks                                  */
-/* --------------------------------------------  */
-
-Hooks.on('renderPause', (app, [html]) => {
-  const img = html.querySelector('img')
-  img.src = 'systems/arcane-arcade-fallout/assets/vaultboy/vaultboy.webp'
-})
-
-Hooks.on('aafohud.skillRoll', async (actorUuid, skill) => {
-  const actor = fromUuidSync(actorUuid)
-  const roll = await new SkillRoll(actor, skill, () => {})
-  roll.render(true)
-})
-
-Hooks.on('aafohud.attackRoll', async (actorUuid, weaponId) => {
-  const actor = fromUuidSync(actorUuid)
-  const weapon = actor.items.get(weaponId)
-  weapon.rollAttack({ advantageMode: 1 })
-})
-
-// AAFO-HUD HOOKS
-Hooks.on('aafohud.toggleEquipArmor', async (actorUuid, itemId) => {
-  const actor = fromUuidSync(actorUuid)
-  const item = actor.items.get(itemId)
-  const cost = item.type == "powerArmor" ? 6 : 3
-  const canAffordAP = actor.applyApCost(cost)
-  if (canAffordAP) {
-    item.update({ 'system.itemEquipped': !item.system.itemEquipped })
-    FalloutZeroArmor.prototype.changeEquipStatus(item)
-  }
-})
-
-Hooks.on('aafohud.toggleEquipWeapon', async (actorUuid, itemId) => {
-  const actor = fromUuidSync(actorUuid)
-  const item = actor.items.get(itemId)
-  const cost = 3
-  const canAffordAP = actor.applyApCost(cost)
-  if (canAffordAP) {
-    item.update({ 'system.itemEquipped': !item.system.itemEquipped })
-  }
-})
-
-Hooks.on('aafohud.reloadWeapon', async (actorUuid, itemId) => {
-  const actor = fromUuidSync(actorUuid)
-  actor.reload(itemId)
-})
-
-Hooks.on('aafohud.useConsumable', async (actorUuid, itemId) => {
-  const actor = fromUuidSync(actorUuid)
-  const item = actor.items.get(itemId)
-  const cost = 4
-  const canAffordAP = actor.applyApCost(cost)
-  if (canAffordAP) {
-    FalloutZeroItem.prototype.toggleEffects(item, item.system.itemEquipped)
-    actor.lowerInventory(itemId)
-  }
-})
+registerHooks()
 
 /* -------------------------------------------- */
 /*  Hotbar Macros                               */
